+++ dataPumper/src/main/java/it/unibz/inf/data_pumper/column_types/StringColumn.java
@@ -143,7 +143,7 @@ public class StringColumn extends OrderedDomainColumn<String> {
 	}
 	
 	@Override
-	public void createNValues(Schema schema, DBMSConnection db, int n) throws ValueUnsetException{
+	public void createNValues(Schema schema, DBMSConnection db, long n) throws ValueUnsetException{
 		
 		List<String> values = new ArrayList<String>();
